Change reader interface wrt Size()
Size() now returns the size of the ROM and the size of any header
present. For most this will be zero however NES and Lynx ROMs can have
headers present. Also clarify that Checksum() will not include the
header when calculating the checksum as this is how the .dat files
reference the ROMs.